+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,7 +3,12 @@
 ## Unreleased
 
 * Added
-  * Added new exaples for cell content and cell selection to `terra-compact-interactive-list`.
+  * Added documentation for exported constants of `terra-table` in the About page.
+  * Added new examples for cell content and cell selection to `terra-compact-interactive-list`.
+
+* Changed
+  * Updated examples and tests for `terra-table` to consume updated onRowSelect callback.
+  * Updated XFC initialization for the Embedded Content Consumer examples.
 
 ## 1.49.0 - (December 1, 2023)