+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,7 +4,20 @@
 
 * Added
   * Added examples and tests for `terra-folder-tree`.
+  * Added documentation for exported constants of `terra-table` in the About page.
+
+* Changed
+  * Updated examples and tests for `terra-table` to consume updated onRowSelect callback.
+  * Updated XFC initialization for the Embedded Content Consumer examples.
+
+## 1.49.0 - (December 1, 2023)
+
+* Added
   * Added new test to `terra-compact-interactive-list` for inconsistent width unit types in props.
+
+* Updated
+  * Updated examples, and tests to show the visual focus indicator on the iframe of the embedded content.
+  * Remove TODO comments that are no longer valid since relative paths work for examples, and tests for Embedded Content Consumers.
 
 ## 1.48.0 - (November 29, 2023)